requirements.txt
fastapi
uvicorn
Pillow
openai
python-multipart
